+++ spring-cli-jpa/src/main/java/samuelmovi/springCliJpa/controller/Controller.java
@@ -70,7 +70,6 @@ public class Controller {
             catch(Exception e){
                 System.out.println("[!!!] Stop messing around");
             }
-
         }
         while (!done);
     }